+++ src/components/chat/Menu/MainMenuContent/index.tsx
@@ -3,13 +3,16 @@ import { useAtomValue } from 'jotai'
 import { useTranslation } from 'next-i18next'
 import { isToday, isYesterday, subDays } from 'date-fns'
 
-import { chatItemsAtom } from '@/atoms'
+import { chatItemsAtom, isSearchActiveAtom, searchQueryNameAtom } from '@/atoms'
 import { type TCategorizedChatItems, type TChatItem } from '@/types'
 import ChatItemCard from '@/components/chat/Menu/MainMenuContent/ChatItemCard'
+import DatabaseOffIcon from '@/components/icons/DatabaseOffIcon'
 
 export default function MainMenuContent() {
-  const chatItemLists =  useAtomValue(chatItemsAtom)
   const {t} = useTranslation('common')
+  const chatItemLists =  useAtomValue(chatItemsAtom)
+  const isSearchActive = useAtomValue(isSearchActiveAtom)
+  const searchQueryName = useAtomValue(searchQueryNameAtom)
 
   // Note: category value must be the same as the key in i18n/common.json
   const renderOrder = useMemo(() => {
@@ -43,6 +46,65 @@ export default function MainMenuContent() {
     }, {})
   }, [chatItemLists, renderOrder])
 
+  if (isSearchActive) {
+    const searchChatItems: TChatItem[] = chatItemLists.filter((chatItem: TChatItem) => chatItem.itemName.includes(searchQueryName))
+
+    if (searchChatItems.length === 0) {
+      return (
+        <div className={'flex-1 overflow-y-auto custom-scrollbar mb-2'}>
+          <DatabaseOffIcon
+            width={48}
+            height={48}
+            className={'flex justify-center mt-8 mb-2'}
+          />
+          <p className={'text-center'}>No data.</p>
+        </div>
+      )
+    }
+
+    return (
+      <div className={'flex-1 overflow-y-auto custom-scrollbar mb-2'}>
+        {/* Star Lists */}
+        <div className={'w-full'}>
+          {
+            searchChatItems
+              .filter((chatItem: TChatItem) => chatItem.isStarred)
+              .sort((a: TChatItem, b: TChatItem) => new Date(b.updatedAt).getTime() - new Date(a.updatedAt).getTime())
+              .map((chatItem: TChatItem) => (
+                <ChatItemCard
+                  key={chatItem.id}
+                  id={chatItem.id}
+                  text={chatItem.itemName}
+                  uuid={chatItem.itemUuid}
+                  isStarred={chatItem.isStarred}
+                />
+              ))
+          }
+        </div>
+
+        <div className={'border-b border-gray-500 mt-1 mb-2'}/>
+
+        {/* No Star Lists */}
+        <div className={'w-full'}>
+          {
+            searchChatItems
+              .filter((chatItem: TChatItem) => !chatItem.isStarred)
+              .sort((a: TChatItem, b: TChatItem) => new Date(b.updatedAt).getTime() - new Date(a.updatedAt).getTime())
+              .map((chatItem: TChatItem) => (
+                <ChatItemCard
+                  key={chatItem.id}
+                  id={chatItem.id}
+                  text={chatItem.itemName}
+                  uuid={chatItem.itemUuid}
+                  isStarred={chatItem.isStarred}
+                />
+              ))
+          }
+        </div>
+      </div>
+    )
+  }
+
   return (
     <div className={'flex-1 overflow-y-auto custom-scrollbar mb-2'}>
       {/* Star Lists */}
@@ -74,7 +136,7 @@ export default function MainMenuContent() {
 
             return (
               <div key={categoryDate} className={'mb-1'}>
-                <p className={'text-xs font-light text-gray-400'}>{t(`chatPage.menu.${categoryDate}`)}</p>
+                <p className={'text-xs font-light text-gray-400 mb-1'}>{t(`chatPage.menu.${categoryDate}`)}</p>
                 <div className={'flex flex-col gap-1'}>
                   {
                     chatItemsList